requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
aiohttp=3.8.5=py39h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow=1.2.3=py39h06a4308_1
arrow-cpp=11.0.0=hda39474_2
async-timeout=4.0.2=py39h06a4308_0
attrs=23.1.0=py39h06a4308_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
bert-score=0.3.13=pypi_0
binaryornot=0.4.4=pyhd3eb1b0_1
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
chardet=4.0.0=py39h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
contourpy=1.1.1=pypi_0
cookiecutter=1.7.3=pyhd3eb1b0_0
cryptography=41.0.3=py39h130f0dd_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
datasets=2.12.0=py39h06a4308_0
dill=0.3.6=py39h06a4308_0
evaluate=0.4.0=py39h06a4308_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.43.1=pypi_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2023.9.2=py39h06a4308_0
gflags=2.2.2=he6710b0_0
glog=0.5.0=h2531618_0
grpc-cpp=1.48.2=h5bf31a4_0
huggingface_hub=0.17.3=py39h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py39h06a4308_0
jinja2-time=0.2.0=pyhd3eb1b0_3
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h568e23c_1
ld_impl_linux-64=2.38=h1181459_1
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=8.2.1=h91b91d3_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.52.0=ha637b67_1
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h0d84882_2
libuv=1.44.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.8.0=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
multidict=6.0.2=py39h5eee18b_0
multiprocess=0.70.14=py39h06a4308_0
ncurses=6.4=h6a678d5_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.0=py39h5f9d8c6_0
numpy-base=1.26.0=py39hb5e798b_0
openssl=1.1.1w=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=23.1=py39h06a4308_0
pandas=2.0.3=py39h1128e8f_0
pillow=10.1.0=pypi_0
pip=23.2.1=py39h06a4308_0
poyo=0.5.0=pyhd3eb1b0_0
pyarrow=11.0.0=py39h468efa6_1
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h5eee18b_1
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=py39h06a4308_0
responses=0.13.3=pyhd3eb1b0_0
scipy=1.11.3=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
text-unidecode=1.3=pyhd3eb1b0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
tqdm=4.65.0=py39hb070fc8_0
transformers=4.22.2=pypi_0
typing-extensions=4.7.1=py39h06a4308_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.16=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.41.2=py39h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py39h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
